+++ CHANGELOG.md
@@ -9,3 +9,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 ### Added
 
 - `runningBehavior` option to CustomEventEmitter
+- `afterAll` hook to CustomEventEmitter